Changed project structure to be more use-case oriented and package-separated.
