+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ui.ExploreArtActivity">
+        <activity android:name=".ui.explore.ExploreArtActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
