+++ app/src/main/java/org/imozerov/streetartview/ui/explore/interfaces/Filterable.kt
@@ -1,4 +1,4 @@
-package org.imozerov.streetartview.ui.interfaces
+package org.imozerov.streetartview.ui.explore.interfaces
 
 /**
  * Created by imozerov on 11.02.16.
