+++ app/src/main/res/layout/activity_explore_art.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.imozerov.streetartview.ui.ExploreArtActivity">
+    tools:context=".ui.explore.ExploreArtActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
