+++ app/src/main/res/layout/fragment_art_list.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.imozerov.streetartview.ui.catalog.ArtListFragment">
+             tools:context=".ui.explore.list.ArtListFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/art_objects_recycler_view"
