+++ app/src/main/res/layout/fragment_art_map.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.imozerov.streetartview.ui.map.ArtMapFragment">
+             tools:context=".ui.explore.map.ArtMapFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
